requirements.txt
alpaca-trade-api
black
click
matplotlib
pandas
pytest
pyyaml
seaborn
sklearn
tensorflow
